Made modifications to allow c&s to run as part of plantech
