+++ src/Dfe.ContentSupport.Web/Views/Content/Index.cshtml
@@ -1,0 +1,6 @@
+﻿@model CsPage
+
+@foreach (var content in @Model.Content)
+{
+    <partial name="_Content" model="@content"/>
+}
