--- src/Dfe.ContentSupport.Web/Views/Home/Index.cshtml
@@ -1,6 +1,0 @@
-﻿@model CsPage
-
-@foreach (var content in @Model.Content)
-{
-    <partial name="_Content" model="@content"/>
-}
