+++ src/Dfe.ContentSupport.Web/Views/_ViewStart.cshtml
@@ -1,3 +1,3 @@
 ﻿@{
-    Layout = "_Layout";
+    Layout = "_CsLayout";
 }
